+++ features/posts/api/index.ts
@@ -17,3 +17,30 @@ export const getAllPostsSlug = () => {
   const markdownFiles = files.filter((file) => extname(file) === ".md");
   return markdownFiles.map((file) => file.replace(".md", ""));
 };
+
+export const getAllPosts = () => {
+  const slugs = getAllPostsSlug();
+  const posts = slugs.map((slug) => {
+    const { data } = getPostBySlug(slug);
+    return {
+      slug,
+      title: data.title,
+      date: data.date,
+      tags: data.tags,
+    } as {
+      slug: string;
+      title: string;
+      date: string;
+      tags?: string[];
+    };
+  });
+
+  return posts;
+};
+
+export const getAllPostsOrderedByDate = () => {
+  const posts = getAllPosts();
+  posts.sort((a, b) => Date.parse(b.date) - Date.parse(a.date));
+
+  return posts;
+};
